pyproject.toml
[tool.poetry]
name = "photo_album"
version = "0.1.0"
description = "Web-based Photo Album written in Python"
authors = ["Alex M <alex@alexos.dev>"]
license = "GNU GPLv3"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.0"
pydantic-settings = "^2.3.4"
jinja2 = "^3.1.4"
jinja2-fragments = "^1.5.0"
pytailwindcss = "^0.2.0"
pillow = "^11.0.0"
google-cloud-storage = "^3.0.0"
PyJWT = "*"
google-cloud-secret-manager = "^2.20.1"
google-crc32c = "*"
requests = "^2.32.3"
authlib = "^1.3.1"
itsdangerous = "^2.2.0"
structlog = "^25.0.0"
# pin gRPC due to MacOS issues in 1.65
grpcio = "1.71.0"
prometheus-fastapi-instrumentator = "^7.0.0"
google-cloud-firestore = "^2.19.0"
uvicorn = "*"
httpx = "^0.28.1"
starlette = "^0.46.2"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.2"
flake8 = "^7.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
